#545 fix problema con checksums al eliminar copias (resta analizar el caso de los .shp)

+++ app/Model/Archivo.php
@@ -65,7 +65,7 @@ class Archivo extends Model
             $checksums[] = $checksum;
             foreach ($shape_files as $key => $shape_file) {
                 // Hay e00 con shapefiles con valor null
-                if ($shape_file != null and $shape_file != ''){
+                if (Storage::exists($shape_file)){ //si no existe el storage para el shapefile no lo tengo en cuenta para el checksum
                   $checksums[] =  md5_file($shape_file);//->getRealPath());
                 }
             }
@@ -724,12 +724,14 @@ class Archivo extends Model
             $c = $nombre_copia . $extension;
             if(Storage::exists($c)) {
                 if(Storage::exists($o)) {
-                    # si existe el archivo original entonces elimino la copia
-                    Log::info("Existe el archivo ".$extension." original en el storage. Se eliminará la copia");
-                    if(Storage::delete($c)){
-                        Log::info('Se borró el archivo: '.  $this->nombre_original . " extension " . $extension);
-                    }else{
-                        Log::error('NO se borró el archivo: '.$this->nombre_original . " extension " . $extension);
+                    # si existe el archivo original y la copia tiene storage distinto entonces elimino la copia
+                    if ($o != $c) { //así no elimino el storage original por error (causando problemas de checksum)
+                        Log::info("Existe el archivo ".$extension." original en el storage. Se eliminará la copia");
+                        if(Storage::delete($c)){
+                            Log::info('Se borró el archivo: '.  $this->nombre_original . " extension " . $extension);
+                        }else{
+                            Log::error('NO se borró el archivo: '.$this->nombre_original . " extension " . $extension);
+                        }
                     }
                 } else {
                     # si no existe entonces el archivo copia reemplaza al original en el storage (son el mismo por lo que no hay problema)
@@ -745,17 +747,19 @@ class Archivo extends Model
         Log::info("Verificando storage");
         $copia = $this;
         if(Storage::exists($original->nombre)) {
-            # si existe el archivo original entonces elimino la copia
-            Log::info("Existe el archivo original en el storage. Se eliminará la copia");
-            if(Storage::delete($copia->nombre)){
-                Log::info('Se borró el archivo: '.$copia->nombre. ' nombre original:'. $copia->nombre_original);
-            }else{
-                Log::error('NO se borró el archivo: '.$copia->nombre. ' nombre original:'.$copia->nombre_original);
+            # si existe el archivo original y la copia tiene storage distinto entonces elimino la copia
+            if ($original->nombre != $copia->nombre) { //así no elimino el storage original por error (causando problemas de checksum)
+                Log::info("Existe el archivo original en el storage. Se eliminará la copia");
+                if(Storage::delete($copia->nombre)){
+                    Log::info('Se borró el archivo: '.$copia->nombre. ' nombre original:'. $copia->nombre_original);
+                }else{
+                    Log::error('NO se borró el archivo: '.$copia->nombre. ' nombre original:'.$copia->nombre_original);
+                }
             }
         } else {
             # si no existe entonces el archivo copia reemplaza al original en el storage (son el mismo por lo que no hay problema)
             Log::error("No existe el archivo original en el storage. Se reutiliza la copia");
-            $original->update(['nombre' => $copia->nombre]);
+            Storage::move($copia->nombre, $original->nombre); 
         }
     }
